# feat: Phase 2 - Advanced Search, Categories, Multi-Select & Performance