… CLAUDE.md
… EPCC_CODE_PHASE2.md
… EPCC_COMMIT_PHASE2.md
… EPCC_PLAN_PYTHON.md
… LESSONS_LEARNED_PHASE2.md
… PERFORMANCE_OPTIMIZATION_SUMMARY.md
… PHASE_2_UX_ACCESSIBILITY_REVIEW.md
… README.md
… SECURITY_REVIEW_PHASE2.md
… SECURITY_SUMMARY.md
… docs/API_REFERENCE.md
… docs/ARCHITECTURE_PHASE2.md
… docs/CONFIGURATION.md
… docs/MIGRATION_PHASE1_TO_PHASE2.md
… docs/PERFORMANCE_BENCHMARKS.md
… docs/PHASE_2_FEATURES.md
… docs/README.md
… docs/TESTING_PHASE2.md
… src/claude_resource_manager/core/search_engine.py
… src/claude_resource_manager/tui/app.py
… src/claude_resource_manager/tui/screens/browser_screen.py
… src/claude_resource_manager/tui/screens/help_screen.py
… src/claude_resource_manager/utils/cache.py
… src/claude_resource_manager/utils/import_profiler.py
… tests/conftest.py
… tests/integration/test_batch_installation.py
… tests/unit/core/test_category_engine.py
… tests/unit/core/test_fuzzy_search.py
… tests/unit/test_performance.py
… tests/unit/tui/test_advanced_ui.py