## Summary

This PR implements **Phase 2** of the Claude Resource Manager, adding advanced search capabilities, category browsing, multi-select batch installation, and comprehensive performance optimizations.

### Key Features

#### 🔍 Advanced Search
- **Fuzzy search** with intelligent scoring and ranking
- Multi-field matching (ID, name, description, tags, categories)
- Sub-20ms search performance on 331+ resources
- Real-time search results with responsive UI

#### 📂 Category Browsing
- Browse resources by category (agents, commands, hooks, templates, mcps)
- Category-specific search and filtering
- Visual category indicators in browser screen
- Seamless category switching with keyboard shortcuts

#### ✅ Multi-Select & Batch Operations
- Space-bar to select/deselect resources
- Visual selection feedback with checkboxes (✓)
- Batch installation of multiple resources
- Progress tracking for batch operations
- "Select all" functionality for efficient workflow

#### ⚡ Performance Optimizations
- **Startup time**: <100ms (catalog loading optimized)
- **Search**: <20ms for fuzzy search, <5ms for exact match
- **Memory**: <100MB with full catalog loaded
- Lazy loading and import profiling
- Smart caching with TTL and size limits

#### ❓ Help System
- Comprehensive help screen (`?` key)
- Keyboard shortcuts reference
- Feature documentation
- Context-sensitive help

### Documentation & Testing

#### 📚 Documentation (8 new docs)
- `API_REFERENCE.md` - Complete API documentation
- `ARCHITECTURE_PHASE2.md` - System architecture and design decisions
- `CONFIGURATION.md` - Configuration guide
- `MIGRATION_PHASE1_TO_PHASE2.md` - Upgrade guide
- `PERFORMANCE_BENCHMARKS.md` - Performance metrics and profiling
- `PHASE_2_FEATURES.md` - Feature documentation
- `TESTING_PHASE2.md` - Testing strategy and coverage
- `LESSONS_LEARNED_PHASE2.md` - Development insights

#### 🧪 Test Coverage
- **367 total tests** (from Phase 1)
- **+200 new tests** for Phase 2 features
- Unit tests for all core modules
- Integration tests for batch workflows
- Performance benchmarks
- >90% coverage on core modules

### Security & UX Reviews

- ✅ **Security review completed** - Path validation, URL sanitization, YAML safety
- ✅ **UX/Accessibility audit** - WCAG compliance, keyboard navigation, visual feedback
- ✅ **Performance profiling** - All targets met or exceeded

### Implementation Details

**Core Modules Added:**
- `core/category_engine.py` - Category management and filtering
- `core/search_engine.py` - Enhanced with fuzzy search
- `utils/cache.py` - Smart caching layer
- `utils/import_profiler.py` - Startup performance monitoring
- `tui/screens/help_screen.py` - Interactive help system

**Enhanced Modules:**
- `tui/screens/browser_screen.py` - Multi-select, categories, visual feedback
- `tui/app.py` - Enhanced navigation and keybindings
- `core/catalog_loader.py` - Performance optimizations

### Breaking Changes

None - fully backward compatible with Phase 1

### Migration Notes

See `docs/MIGRATION_PHASE1_TO_PHASE2.md` for upgrade instructions

### Test Plan

- [x] Run full test suite (567+ tests passing)
- [x] Verify fuzzy search performance (<20ms)
- [x] Test batch installation with 5+ resources
- [x] Validate category filtering across all types
- [x] Confirm help screen displays correctly
- [x] Check startup time (<100ms)
- [x] Security review passed
- [x] UX/accessibility audit passed

### Workflow Improvements

This PR also includes updates to `CLAUDE.md`:
- **Branch naming convention**: `{user}/{feature-description}` format
- **PR requirement**: All changes must go through PR review
- **Git workflow**: Step-by-step guide for feature development

---

**Stats:** 35 files changed, 16,803 insertions(+), 132 deletions(-)

🤖 Generated with [Claude Code](https://claude.com/claude-code)